test: organize responses api unit tests
